Implementing an Ember application for Presto
The application consists of a main controller, which observes three
aspects of the system. First is the screen. When the screen size changes
height, the main controller will re-arrange the text (if any) is being
displayed. Second is the font. There is no hook to keep the font model
updated. Third is text area. When the text area has its value changed,
the main controller will re-arrange the new value.

Jasmine has been added for testing, and a runner can be found in the
spec directory, under jasmine. Viewing this in the browser will execute
the specs.